Create getCountryCodeFromURL function
